requirements.txt
pandas>=2.1.0
numpy>=1.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
loguru>=0.7.2
rich>=13.7.0
httpx>=0.26.0
pybit>=5.7.0
optuna>=3.5.0
plotly>=5.18.0
streamlit>=1.29.0
pyarrow>=14.0.0
typer>=0.9.0
pyyaml>=6.0.1
scipy>=1.12.0
pytest>=7.4.0
kaleido>=0.2.1
tqdm>=4.66.0

